Add a test server and client.
This PR creates two new executables, 'server' and 'client'.

The first starts a server that mimicks the behaviour of the consensus
and networking layer.

The second is a version of radicle with two new primops, 'send!' and
'receive!'. The first allows submitting new expressions to a chain. The
takes a dictionary of chains, possibly with some already-known state and
expressions, and returns a new dictionary with updated state and
expressions.

The directory structure has also been updated to make more sense now
that the REPL isn't the only executable (and now that the radicle files
are not specific to the REPL).