Add a recursive version of `define`.
Add a form `(define-rec <symbol> <def>)`. This form evaluates `<def>`, and if
this results in a closure whose body refers to <symbol>, then the result of the
`define-rec` is to bind <symbol> to a recursive version of the closure. If the
result of evaluating `<def>` is not a closure, then the result is an error.