Improve chain API
We stop using the`Show` and `Read` instances to serialise Radicle values
for the plaintext MIME type. Instead we render the radicle values as
code and parse them as code.

This also changes the contract of the `since` endpoint. Now we return
the expresssions as a list of Radicle values.