Build a working image of radicle-server
We add the capability to build a working `radicle-server` image locally
and on CI. We push the image to
`eu.gcr.io/opensourcecoin/radicle-server` on CI.

The previous setup using Stack’s [`image` configuration][1] did not work
because the image requires `libpq`. Now we add the
`images/radicle-server` folder that holds the `Dockerfile` and scripts
to build the image locally and on CI.

Note that for local builds with the `build.sh` script we tell Stack to
build the binaries inside a Docker conatiner (using `stack build
--docker`). This ensures that the produced binary is compatible with the
distribution in the base image.

[1]: https://docs.haskellstack.org/en/latest/GUIDE/#docker