Use a custom stack snapshot and improve CI cache
We now use a custom stack snapshot instead of specifying `extra-deps` in
`stack.yaml` and tweak the CI caching scripts.

Before the CI build caching suffered from a couple of problems:

* A cache for `STACK_ROOT` would always be uploaded after the build.
  (I don’t know why.)
* We cached the build artifacts in `.stack-work` keyed by the has of
  `stack.yaml`. This resulted in many builds sharing building artifacts
  which could potentially result in bugs.

With this commit we stop caching build artifacts and improve the caching
strategy for `STACK_ROOT`.

By using a snapshot the artifacts for the extra packages (e.g. `cbor`)
are build in `STACK_ROOT` and cached. (Including the extra packages in
the cache was the initial reason for caching `.stack-work`)

The main cache ID for `STACK_ROOT` is determined by the hash of
`snapshot.yaml` and `package.yaml`. (Ideally the cache ID should be the
hash of the output of `stack ls dependencies`, but this requires some
further modifications.)

In addition we have a `STACK_ROOT` cache for the master branch. This
cache is always uploaded for a master build. If a cache with a specific
hash does not exist we load the master cache.