Tag radicle-server images with commit sha
In addition to pushing the `radicle-server` image with the `latest` tag
we push the same image with the date and short Git commit SHA as the
tag. An example looks like this:

    b2018.12.06-a76a52f

This gives us both an indicator of then the image was build and what the
commit SHA is.