Sys args, open-file, and unify util scripts (#296)
Adds:

* (get-args!)
* (open-file!)
* (write-file!)
* (edit-in-editor!)

A little work to make shebangs possible.

Moves list-issues and new-issue to a single file, and make it a pure radicle file.

In theory `read-file!` no longer has to be a primop. But testing would be quite complicated because of the way files are mocked (which is very relevant to load! and so to everything). I'd open a separate issue for that.