Use machine backend type to make primfn names
We rename the primitive functions provided by the `eval-server` machine
backend to `machine/eval-server/update!` `machine/eval-server/get-log!`.

This allows us to restructure the Machine Backend interface. We omit the
primitive function name and the documentation for the generated
primitives. They all implement the machine backend interface so the
documentation is the same.

To change as little code as possible we define the old `send!` and
`receive!` names in the `prelude/chain` module. These will be extended
when the IPFS backend is implemented.