Save build cache even on failure
We want to save the build cache for dependencies even if the project
code fails to build or the tests fail. Before a test failure would
prevent the build cache from being uploaded. This meant that the next
build that fixed the tests would need to rebuild all the new
dependencies.