Isolate e2e tests
This commit improves the isolation of the end-to-end tests for the
counter app by running the commands with a clean environment in a
temporary directory.

This also necessitates some refactoring. We introduce the
`Test.E2ESupport` module which provides the `TestM` monad. All generic
code for tests is moved there.