daemon: Wait for IPFS daemon
With systemd the Radicle daemon is started very quickly after the IPFS
daemon has started. Therefore the IPFS daemon might not be ready
yet when the Radicle daemon tries to load the machines. This results in
errors.

To prevent this the Radicle daemon waits until the IPFS daemon is ready
by repeatedly trying to connect to its `version` endpoint.