Make pattern matching work in prelude modules
This fixes a bug where one of the prelude modules uses extended pattern
matching (as defined in `match-pat`) without importing `prelude/patterns`.

For example one of the prelude modules may contain the code

    (def foo
      (match bar
        :foo #t
        _    #f))

If `match-pat` from `prelude/patterns` is not in the global environment
when this module is evaluated then the `match` special form will call
the matcher `:foo` as a function which will throw an error.