feat: New show command (#638)
BREAKING CHANGE: This changes the structure of the message that is
signed before sending it to a machine. Machines set up before this
commit will not accept the new signatures.

`show-unbound` does not add formatting to the requests payload, to
make signing request consistently to work also in other languages.