feat: Transactor functions.
Transactor re-definition replaces eval re-difinition.

This is a much simpler model. At first tx is set to the identity function, that is, the machine accepts all input expressions as is, and evaluates them.

One can then redefine tx, to restrict the inputs to the machine. If tx throws an exception, then the input is not accepted, otherwise it can perform effects itself (e.g. mutate some refs), and then return an expression for evaluation on the machine.

See https://github.com/radicle-dev/transparent-machines-doc for more information.

BREAKING CHANGE: This totally changes how machines restrict their inputs, so is a major breaking change.